requirements.txt
Flask==0.12 
flask-restful==0.3.5
docker==4.2.2
paho-mqtt==1.5.0	
pandas==1.0.5